requirements.txt
openai>=1.12.0
duckdb>=0.10.0
chromadb>=0.4.22
lxml>=5.1.0
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
matplotlib>=3.8.0
typer>=0.9.0
python-dotenv>=1.0.0
rich>=13.7.0
